backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
graphene==3.3
strawberry-graphql==0.219.2
neo4j==5.15.0
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-dotenv==1.0.1
pydantic==2.6.1
pytest==8.0.0
httpx==0.26.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9 